requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
datasets==3.0.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docopt==0.6.2
einops==0.8.0
fastapi==0.115.0
filelock==3.16.1
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
gguf==0.10.0
grip==4.6.2
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.5.0
interegular==0.3.3
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lark==1.2.2
llvmlite==0.43.0
lm-format-enforcer==0.10.6
Markdown==3.7
MarkupSafe==2.1.5
mistral_common==1.4.4
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
mss==9.0.2
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-ml-py==12.560.30
openai==1.51.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outlines==0.0.46
packaging==24.1
pandas==2.2.3
partial-json-parser==0.2.1.1.post4
path-and-address==2.0.1
pillow==10.4.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
protobuf==5.28.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
PyAutoGUI==0.9.54
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.37.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
rubicon-objc==0.4.9
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
tiktoken==0.7.0
tokenizers==0.20.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
vllm==0.6.2
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.4
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.2
